+++ lascli/parser/models.py
@@ -1,6 +1,9 @@
+import textwrap
+from argparse import RawTextHelpFormatter
+
 from las import Client
 
-from lascli.util import NotProvided, nullable, json_path, nullable_json_path
+from lascli.util import NotProvided, nullable, json_path, nullable_json_path, json_or_json_path
 
 
 def create_model(las_client: Client, field_config, **optional_args):
@@ -112,7 +115,7 @@ def create_models_parser(subparsers):
     get_parser.add_argument('model_id')
     get_parser.set_defaults(cmd=get_model)
 
-    update_parser = subparsers.add_parser('update')
+    update_parser = subparsers.add_parser('update', formatter_class=RawTextHelpFormatter)
     update_parser.add_argument('model_id')
     update_parser.add_argument(
         '--field-config',
@@ -120,18 +123,12 @@ def create_models_parser(subparsers):
         type=json_path,
         help='path to configuration of the fields that the model will predict',
     )
-    update_parser.add_argument(
-        '--preprocess-config',
-        '-p',
-        type=json_path,
-        help='path to configuration of the step before the prediction',
-    )
     update_parser.add_argument('--name', type=nullable(str), default=NotProvided)
     update_parser.add_argument('--description', type=nullable(str), default=NotProvided)
     update_parser.add_argument(
         '--metadata',
         type=nullable_json_path,
-        help='path to json file with whatever you need, maximum limit 4kB',
+        help='path to json file with custom metadata, maximum limit 4kB',
         default=NotProvided,
     )
     update_parser.add_argument(
@@ -140,6 +137,31 @@ def create_models_parser(subparsers):
         default=NotProvided,
         help='Use training_id for model inference in POST /predictions. Specify "null" to make model inactive.',
     )
+    update_parser.add_argument('--preprocess-config', type=json_or_json_path, help=textwrap.dedent('''
+        Path or inline JSON with the pre processing configuration for predictions made by this model
+        {
+            "autoRotate": true | false,                 (required)
+            "imageQuality": "LOW" | "HIGH",             (required)
+            "maxPages": 1 - 3                           (required)
+        }
+        Examples:
+        {"imageQuality": "HIGH", "autoRotate": false, "maxPages": 3}
+        {"imageQuality": "LOW", "autoRotate": true, "maxPages": 1}
+    '''))
+    update_parser.add_argument('--postprocess-config', type=json_or_json_path, help=textwrap.dedent('''
+        Path or inline JSON with the post processing configuration for predictions made by this model
+        {
+            "strategy": "BEST_FIRST" | "BEST_N_PAGES",  (required)
+            "parameters": {                             (required if strategy=BEST_N_PAGES, omit otherwise)
+                "n": int,                               (required if strategy=BEST_N_PAGES, omit otherwise)
+                "collapse": true | false (default)      (optional if strategy=BEST_N_PAGES, omit otherwise)
+            }
+        }
+        Examples:
+        {"strategy": "BEST_FIRST"}
+        {"strategy": "BEST_N_PAGES", "parameters": {"n": 3}}
+        {"strategy": "BEST_N_PAGES", "parameters": {"n": 3, "collapse": true}}
+    '''))
     update_parser.set_defaults(cmd=update_model)
 
     delete_parser = subparsers.add_parser('delete')